requirements.txt
Flask
scikit-learn
tensorflow
pandas 
numpy
matplotlib
seaborn
pytesseract
streamlit
fitz 
yake
pillow
spacy
google-generativeai 
python-dotenv 
pdf2image
